+++ src/file_grepper.h
@@ -1,6 +1,6 @@
 /*
  Sniffle
- Copyright 2018 Peter Pearson.
+ Copyright 2018-2019 Peter Pearson.
 
  Licensed under the Apache License, Version 2.0 (the "License");
  You may not use this file except in compliance with the License.
@@ -74,6 +74,9 @@ private:
 	// match items
 	MatchType			m_matchType;
 	std::vector<std::string>	m_aMatchItems;
+	
+	bool				m_shortCircuit;
+	std::string			m_shortCircuitString;
 };
 
 #endif // FILE_GREPPER_H